requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

scipy==1.7.0
optuna==2.8.0
imbalanced_learn==0.8.0
lightgbm==3.2.1
matplotlib==3.4.2
pandas==1.3.0
joblib==1.0.1
Boruta==0.3
seaborn==0.11.1
click==7.1.2
numpy==1.21.0
auto_mix_prep==0.2.0
imblearn==0.0
python-dotenv==0.19.0
scikit_learn==0.24.2
